requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
python-jose
passlib[bcrypt]
apscheduler
email-validator
python-dotenv
